tests/wpt/web-platform-tests/svg/painting/parsing/color-rendering-valid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing color-rendering with valid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#ColorRenderingProperty"/>
    <h:meta name="assert" content="color-rendering supports the full grammar 'auto | optimizeSpeed | optimizeQuality'."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("color-rendering", "auto");
test_valid_value("color-rendering", "optimizespeed");
test_valid_value("color-rendering", "optimizequality");

  ]]></script>
</svg>
